// src/types/ClusterComplianceReportSpecComplianceControls.ts
/**
 * 
 * 
 *
 * The version of the OpenAPI document: 
 * Contact Email: 
 * License: 
 *
 * NOTE: This file is auto generated by crdtotypes (https://github.com/yaacov/crdtoapi/).
 * https://github.com/yaacov/crdtoapi/README.crdtotypes
 */

import { ClusterComplianceReportSpecComplianceControlsChecks } from './ClusterComplianceReportSpecComplianceControlsChecks';
import { ClusterComplianceReportSpecComplianceControlsCommands } from './ClusterComplianceReportSpecComplianceControlsCommands';

/**
 * Control represent the cps controls data and mapping checks
 *
 * @export
 */
export interface ClusterComplianceReportSpecComplianceControls {
  /** checks
   * SpecCheck represent the scanner who perform the control check
   *
   * @required {false}
   */
  checks?: ClusterComplianceReportSpecComplianceControlsChecks[];
  /** commands
   * Commands represent the commands to be executed by the node-collector
   *
   * @required {false}
   */
  commands?: ClusterComplianceReportSpecComplianceControlsCommands[];
  /** defaultStatus
   * define the default value for check status in case resource not found
   *
   * @required {false}
   * @originalType {string}
   */
  defaultStatus?: 'PASS' | 'WARN' | 'FAIL';
  /** description
   *
   * @required {false}
   */
  description?: string;
  /** id
   * id define the control check id
   *
   * @required {true}
   */
  id: string;
  /** name
   *
   * @required {true}
   */
  name: string;
  /** severity
   * define the severity of the control
   *
   * @required {true}
   * @originalType {string}
   */
  severity: 'CRITICAL' | 'HIGH' | 'MEDIUM' | 'LOW' | 'UNKNOWN';
}